* develop:
  #160 fix: namber of table rows in 'DPU instances' from settings
  Fixing #125 Data Lost by warning user and terminating pipeline
  Local rdf enabled query browser
  bugfix: wrong equals - returned always false
  bugfix: wrong equals - returned always false
  remove mff thirdparty repository
  528 bugfix importing dpu with different version caused error, fixed bug when child dpu was created after importing dpu from the same zip twice

+++ commons-app/src/main/java/cz/cuni/mff/xrg/odcs/commons/app/pipeline/DbExecutionImpl.java
@@ -112,6 +112,6 @@ class DbExecutionImpl extends DbAccessBase<PipelineExecution> implements DbExecu
     	TypedQuery<Long> query = createCountTypedQuery(stringQuery);
     	query.setParameter("ids", ids);
     	Long number = (Long) query.getSingleResult();
-    	return !number.equals(ids.size());
+    	return !number.equals((long)ids.size());
     }
 }
